optimize variable import
